requirements.txt
Flask==2.0.0
Jinja2==3.0.0
python-dotenv==0.17.1
mysql-connector-python==8.0.25
matplotlib==3.4.2
networkx==2.5.1